requirements.txt
aiofiles==23.2.1
aiosqlite==0.20.0
asyncua==1.1.0
cffi==1.16.0
cryptography==42.0.5
numpy==1.26.4
pandas==2.2.1
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyX==0.16
scapy==2.5.0
six==1.16.0
sortedcontainers==2.4.0
typing_extensions==4.11.0
tzdata==2024.1
python-dotenv==1.0.1